pyproject.toml
[project]
name = "mygifcompressor"
version = "0.1.0"
description = "Compress GIF files under 1MB while preserving original dimensions"
readme = "README.md"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.11,<4.0"
dependencies = [
    "Pillow>=10.4.0,<11.0.0",
]

[project.scripts]
mygifcompressor = "mygifcompressor.cli:main"

[tool.poetry]
packages = [{ include = "mygifcompressor", from = "src" }]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.6.2"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
